move route to new fixture
